remove local openzeppelin

+++ contracts/NFT.sol
@@ -2,22 +2,22 @@
 
 pragma solidity ^0.8.4;
 
-import "./@openzeppelin/contracts/token/ERC721/ERC721.sol";
-import "./@openzeppelin/contracts/token/ERC721/extensions/ERC721URIStorage.sol";
-import "./@openzeppelin/contracts/utils/Counters.sol";
+import "@openzeppelin/contracts/token/ERC721/ERC721.sol";
+import "@openzeppelin/contracts/token/ERC721/extensions/ERC721URIStorage.sol";
+import "@openzeppelin/contracts/utils/Counters.sol";
 
 contract NFT is ERC721URIStorage {
     using Counters for Counters.Counter;
     Counters.Counter private _tokenIds;
     address contractAddress;
 
-    constructor(address marketPlaceAddress) ERC721("Metaverse Tokens", "METT"){
+    constructor(address marketPlaceAddress) ERC721("Metaverse Tokens", "METT") {
         contractAddress = marketPlaceAddress;
     }
 
-    function createToken(string memory tokenURI) public returns (uint){
+    function createToken(string memory tokenURI) public returns (uint256) {
         _tokenIds.increment();
-        uint newItemId = _tokenIds.current();
+        uint256 newItemId = _tokenIds.current();
         _mint(msg.sender, newItemId);
         _setTokenURI(newItemId, tokenURI);
         setApprovalForAll(contractAddress, true);
